artifact

+++ python/tvm/auto_scheduler/measure.py
@@ -574,6 +574,9 @@ class LocalRPCMeasureContext:
             cooldown_interval,
             enable_cpu_cache_flush,
         )
+        self.key = device_key
+        self.host = "127.0.0.1"
+        self.port = self.tracker.port
         # Wait for the processes to start
         time.sleep(0.5)
 
@@ -1036,6 +1039,7 @@ def _timed_rpc_run(
     cooldown_interval,
     enable_cpu_cache_flush,
     verbose,
+    device_id,
 ):
     # pylint: disable=import-outside-toplevel
     from .search_task import get_task_input_buffer  # lazily import to avoid recursive dependency
@@ -1050,7 +1054,7 @@ def _timed_rpc_run(
         remote = request_remote(key, host, port, priority, timeout)
         remote.upload(build_res.filename)
         func = remote.load_module(os.path.split(build_res.filename)[1])
-        dev = remote.device(str(inp.task.target), 0)
+        dev = remote.device(str(inp.task.target), device_id)
         # Limitation:
         # We can not get PackFunction directly in the remote mode as it is wrapped
         # under the std::function. We could lift the restriction later once we fold
@@ -1152,7 +1156,7 @@ def _rpc_run_worker(args):
     res : MeasureResult
         The measure result of this Runner thread.
     """
-    _, build_res, _, _, _, _, timeout, _, _, _, _, _, verbose = args
+    _, build_res, _, _, _, _, timeout, _, _, _, _, _, verbose, _ = args
     if build_res.error_no != MeasureErrorNo.NO_ERROR:
         return (
             (MAX_FLOAT,),
@@ -1186,6 +1190,39 @@ def _rpc_run_worker(args):
 
     return res
 
+
+def rpc_runner_run_parallel(args):
+    n_parallel, inputs, build_results, key, host, port, priority, timeout, number, repeat, min_repeat_ms, cooldown_interval, enable_cpu_cache_flush, verbose, device_id = args
+
+    # This pool is not doing computationally intensive work, so we can use threads
+    pool = multiprocessing.pool.ThreadPool(n_parallel)
+    tuple_res = pool.map(
+        _rpc_run_worker,
+        [
+            (
+                inp.serialize(),
+                build_res,
+                key,
+                host,
+                port,
+                priority,
+                timeout,
+                number,
+                repeat,
+                min_repeat_ms,
+                cooldown_interval,
+                enable_cpu_cache_flush,
+                verbose,
+                device_id,
+            )
+            for inp, build_res in zip(inputs, build_results)
+        ],
+    )
+    pool.terminate()
+    pool.join()
+    del pool
+
+    return tuple_res
 
 @tvm._ffi.register_func("auto_scheduler.rpc_runner.run")
 def rpc_runner_run(
@@ -1258,32 +1295,35 @@ def rpc_runner_run(
         The measure results of these MeasureInputs.
     """
     assert len(inputs) == len(build_results), "Measure input size should be equal to build results"
-    # This pool is not doing computationally intensive work, so we can use threads
-    pool = multiprocessing.pool.ThreadPool(n_parallel)
-    tuple_res = pool.map(
-        _rpc_run_worker,
-        [
-            (
-                inp.serialize(),
-                build_res,
-                key,
-                host,
-                port,
-                priority,
-                timeout,
-                number,
-                repeat,
-                min_repeat_ms,
-                cooldown_interval,
-                enable_cpu_cache_flush,
-                verbose,
-            )
-            for inp, build_res in zip(inputs, build_results)
-        ],
-    )
-    pool.terminate()
-    pool.join()
-    del pool
+
+    device_number = int(os.environ.get('TVM_DEVICE_NUMBER', '1'))
+    run_per_device = len(inputs) // device_number
+    end_index = 0
+    args_list = []
+    for i in range(device_number):
+        if i < len(inputs) % device_number:
+            run_times = run_per_device + 1
+        else:
+            run_times = run_per_device
+        if i != 0:
+            key = os.environ.get(f"TVM_RPC_KEY-{i}", "-1")
+            host = os.environ.get(f"TVM_RPC_HOST-{i}", "-1")
+            port = int(os.environ.get(f"TVM_RPC_PORT-{i}", "-1"))
+            assert '-1' not in [key, host] and -1 != port
+        args_list.append((n_parallel, inputs[end_index: end_index + run_times], build_results[end_index: end_index + run_times], key, host, port, priority, timeout, number, repeat, min_repeat_ms, cooldown_interval, enable_cpu_cache_flush, verbose, i))
+        end_index += run_times
+
+    if device_number > 1:
+        pool = multiprocessing.pool.ThreadPool(device_number)
+        tuple_res_list = pool.map(rpc_runner_run_parallel, args_list)
+        pool.terminate()
+        pool.join()
+        del pool
+        tuple_res = []
+        for entry in tuple_res_list:
+            tuple_res += entry
+    else:
+        tuple_res = rpc_runner_run_parallel(args_list[0])
 
     results = []
     for res in tuple_res:
